+++ lib/src/models/hive_registration_model.dart
@@ -1,0 +1,15 @@
+import 'package:hive_flutter/adapters.dart';
+
+class HiveRegistrationModel<T> {
+  TypeAdapter<T> adapter;
+  int typeId;
+  bool internal;
+  bool override;
+
+  HiveRegistrationModel({
+    required this.adapter,
+    required this.typeId,
+    this.internal = false,
+    this.override = false,
+  });
+}